Some missing NPE checks

+++ src/main/java/hr/prism/board/service/ResourcePatchService.java
@@ -17,16 +17,16 @@ import java.util.*;
 @Service
 @Transactional
 public class ResourcePatchService {
-    
+
     @Inject
     private DocumentService documentService;
-    
+
     @Inject
     private LocationService locationService;
-    
+
     @Inject
     private ResourceService resourceService;
-    
+
     public void patchName(Resource resource, Optional<String> newValueOptional, ExceptionCode required, ExceptionCode unique) {
         if (newValueOptional != null) {
             String oldValue = resource.getName();
@@ -36,7 +36,7 @@ public class ResourcePatchService {
                     if (unique != null) {
                         resourceService.validateUniqueName(resource.getScope(), resource.getId(), resource.getParent(), newValue, unique);
                     }
-                    
+
                     patchProperty(resource, "name", resource::setName, oldValue, newValue);
                 }
             } else if (required != null) {
@@ -46,7 +46,7 @@ public class ResourcePatchService {
             }
         }
     }
-    
+
     public void patchHandle(Resource resource, Optional<String> newValueOptional, ExceptionCode required, ExceptionCode unique) {
         if (newValueOptional != null) {
             String oldValue = resource.getHandle();
@@ -56,11 +56,11 @@ public class ResourcePatchService {
                 if (!Objects.equals(resource, parent)) {
                     newValue = parent.getHandle() + "/" + newValue;
                 }
-                
+
                 if (unique != null) {
                     resourceService.validateUniqueHandle(resource, newValue, unique);
                 }
-                
+
                 patchHandle(resource, oldValue, newValue);
             } else if (required != null) {
                 throw new ApiException(required);
@@ -69,32 +69,33 @@ public class ResourcePatchService {
             }
         }
     }
-    
+
     public <T> void patchProperty(Resource resource, String property, Getter<T> getter, Setter<T> setter, Optional<T> newValueOptional) {
         patchProperty(resource, property, getter, setter, newValueOptional, null, null);
     }
-    
+
     public <T> void patchProperty(Resource resource, String property, Getter<T> getter, Setter<T> setter, Optional<T> newValueOptional, ExceptionCode required) {
         patchProperty(resource, property, getter, setter, newValueOptional, required, null);
     }
-    
+
     public <T> void patchProperty(Resource resource, String property, Getter<T> getter, Setter<T> setter, Optional<T> newValueOptional, Runnable after) {
         patchProperty(resource, property, getter, setter, newValueOptional, null, after);
     }
-    
+
     public void patchDocument(Resource resource, String property, Getter<Document> getter, Setter<Document> setter, Optional<DocumentDTO> newValueOptional) {
         patchDocument(resource, property, getter, setter, newValueOptional, null);
     }
-    
+
     public void patchDocument(Resource resource, String property, Getter<Document> getter, Setter<Document> setter, Optional<DocumentDTO> newValueOptional, Runnable after) {
         if (newValueOptional != null) {
             Document oldValue = getter.get();
             if (newValueOptional.isPresent()) {
+                String oldValueCloudinaryId = oldValue != null ? oldValue.getCloudinaryId() : null;
                 DocumentDTO newValue = newValueOptional.get();
-                if (!Objects.equals(oldValue.getCloudinaryId(), newValue.getCloudinaryId())) {
+                if (!Objects.equals(oldValueCloudinaryId, newValue.getCloudinaryId())) {
                     patchDocument(resource, property, setter, oldValue, newValue);
                 }
-                
+
                 if (after != null) {
                     after.run();
                 }
@@ -103,7 +104,7 @@ public class ResourcePatchService {
             }
         }
     }
-    
+
     public void patchLocation(Resource resource, Optional<LocationDTO> newValueOptional, ExceptionCode required) {
         if (newValueOptional != null) {
             Location oldValue = resource.getLocation();
@@ -119,7 +120,7 @@ public class ResourcePatchService {
             }
         }
     }
-    
+
     public void patchCategories(Resource resource, CategoryType categoryType, Optional<List<String>> newValuesOptional) {
         if (newValuesOptional != null) {
             List<String> oldValues = resourceService.getCategories(resource, categoryType);
@@ -134,7 +135,7 @@ public class ResourcePatchService {
             }
         }
     }
-    
+
     private <T> void patchProperty(Resource resource, String property, Getter<T> getter, Setter<T> setter, Optional<T> newValueOptional, ExceptionCode required, Runnable after) {
         if (newValueOptional != null) {
             T oldValue = getter.get();
@@ -143,7 +144,7 @@ public class ResourcePatchService {
                 if (!Objects.equals(oldValue, newValue)) {
                     patchProperty(resource, property, setter, oldValue, newValue);
                 }
-                
+
                 if (after != null) {
                     after.run();
                 }
@@ -154,39 +155,43 @@ public class ResourcePatchService {
             }
         }
     }
-    
+
     private <T> void patchProperty(Resource resource, String property, Setter<T> setter, T oldValue, T newValue) {
         setter.set(newValue);
         resource.getChangeList().put(property, oldValue, newValue);
     }
-    
+
     private void patchHandle(Resource resource, String oldValue, String newValue) {
         resourceService.updateHandle(resource, newValue);
         resource.getChangeList().put("handle", oldValue, newValue);
     }
-    
+
     private void patchDocument(Resource resource, String property, Setter<Document> setter, Document oldValue, DocumentDTO newValue) {
         setter.set(documentService.getOrCreateDocument(newValue));
         patchProperty(resource, property, setter, oldValue, documentService.getOrCreateDocument(newValue));
-        documentService.deleteDocument(oldValue);
+        if (oldValue != null) {
+            documentService.deleteDocument(oldValue);
+        }
     }
-    
+
     private void patchLocation(Resource resource, Location oldValue, LocationDTO newValue) {
         patchProperty(resource, "location", resource::setLocation, oldValue, locationService.getOrCreateLocation(newValue));
-        locationService.deleteLocation(oldValue);
+        if (oldValue != null) {
+            locationService.deleteLocation(oldValue);
+        }
     }
-    
+
     private void patchCategories(Resource resource, CategoryType categoryType, List<String> oldValues, List<String> newValues) {
         resourceService.updateCategories(resource, categoryType, newValues);
         resource.getChangeList().put(categoryType.name().toLowerCase() + "Categories", oldValues, newValues);
     }
-    
+
     public interface Getter<T> {
         T get();
     }
-    
+
     public interface Setter<T> {
         void set(T value);
     }
-    
+
 }
